portfolio_dashboard/dashboard/requirements.txt
streamlit>=1.30
pandas>=2.0
numpy>=1.24
yfinance>=0.2.0
plotly>=5.18
feedparser>=6.0
schedule>=1.2
openai>=1.0
pyyaml>=6.0
pandas-ta>=0.3.14b
ib_async>=1.0.0
